requirements.txt
fastapi
uvicorn
pandas
numpy
pillow
matplotlib
sqlalchemy
pytest
httpx
pydantic